requirements.txt
deepgram-sdk>=3.0.0
pynput>=1.7.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
asyncio
